Added guard for application runner, when Kill called and process has already exited

+++ SimpleCalculator.Tests.Acceptance/ApplicationRunner.cs
@@ -25,7 +25,10 @@ namespace SimpleCalculator.Tests.Acceptance
 
         ~ApplicationRunner()
         {
-            _process.Kill();
+            if (!_process.HasExited)
+            {
+                _process.Kill();
+            }
         }
 
         internal void Displays(string expectedText)
